+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 - Added PixFileFormat.Webp and PixWebpSaveParams
 - Added more known image file extensions
 - Added PixTiffSaveParams for choosing the compression scheme when saving TIFF images
+- Added PixExrSaveParams for choosing the compression scheme when saving EXR images
 
 ### 5.3.10
 - [IO] Added SymMap constructor overload